--- src/modules/auth/index.ts
@@ -1,3 +1,0 @@
-export * from './service/auth.service';
-export * from './auth.module';
-export * from './guard/jwt-guard';
