--- src/modules/main/app.controller.ts
@@ -1,16 +1,0 @@
-import { Get, Controller, UseGuards } from '@nestjs/common';
-import { AuthGuard } from '@nestjs/passport';
-import { ApiBearerAuth } from '@nestjs/swagger';
-import { AppService } from './app.service';
-
-@ApiBearerAuth()
-@Controller()
-export class AppController {
-  constructor(private readonly appService: AppService) {}
-
-  @Get()
-  @UseGuards(AuthGuard())
-  root(): string {
-    return this.appService.root();
-  }
-}
