--- src/modules/main/app.service.ts
@@ -1,11 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { ConfigService } from './../config';
-
-@Injectable()
-export class AppService {
-  constructor(private config: ConfigService) {}
-
-  root(): string {
-    return this.config.get('APP_URL');
-  }
-}
